+++ requests_auth/__init__.py
@@ -5,10 +5,12 @@ from requests_auth._authentication import (
     NTLM,
     SupportMultiAuth,
 )
+from requests_auth._oauth2.browser import DisplaySettings
 from requests_auth._oauth2.common import OAuth2
 from requests_auth._oauth2.authorization_code import (
     OAuth2AuthorizationCode,
     OktaAuthorizationCode,
+    WakaTimeAuthorizationCode,
 )
 from requests_auth._oauth2.authorization_code_pkce import (
     OAuth2AuthorizationCodePKCE,
@@ -46,6 +48,7 @@ __all__ = [
     "HeaderApiKey",
     "QueryApiKey",
     "OAuth2",
+    "DisplaySettings",
     "OAuth2AuthorizationCodePKCE",
     "OktaAuthorizationCodePKCE",
     "OAuth2Implicit",
@@ -55,6 +58,7 @@ __all__ = [
     "AzureActiveDirectoryImplicitIdToken",
     "OAuth2AuthorizationCode",
     "OktaAuthorizationCode",
+    "WakaTimeAuthorizationCode",
     "OAuth2ClientCredentials",
     "OktaClientCredentials",
     "OAuth2ResourceOwnerPasswordCredentials",