+++ requests_auth/_oauth2/common.py
@@ -4,6 +4,7 @@ from typing import Optional
 import requests.auth
 
 from requests_auth._errors import InvalidGrantRequest, GrantNotProvided
+from requests_auth._oauth2.browser import DisplaySettings
 from requests_auth._oauth2.tokens import TokenMemoryCache
 
 
@@ -33,6 +34,17 @@ def _get_query_parameter(url: str, param_name: str) -> Optional[str]:
     return all_values[0] if all_values else None
 
 
+def _content_from_response(response: requests.Response) -> dict:
+    content_type = response.headers.get("content-type")
+    if content_type == "text/html; charset=utf-8":
+        return {
+            key_values[0]: key_values[1]
+            for key_value in response.text.split("&")
+            if (key_values := key_value.split("=")) and len(key_values) == 2
+        }
+    return response.json()
+
+
 def request_new_grant_with_post(
     url: str, data, grant_name: str, timeout: float, session: requests.Session
 ) -> (str, int, str):
@@ -42,7 +54,7 @@ def request_new_grant_with_post(
             # As described in https://tools.ietf.org/html/rfc6749#section-5.2
             raise InvalidGrantRequest(response)
 
-        content = response.json()
+        content = _content_from_response(response)
     token = content.get(grant_name)
     if not token:
         raise GrantNotProvided(grant_name, content)
@@ -51,3 +63,4 @@ def request_new_grant_with_post(
 
 class OAuth2:
     token_cache = TokenMemoryCache()
+    display = DisplaySettings()